+++ extensions/common-directory/src/send-file-to-directory.tsx
@@ -16,7 +16,7 @@ import {
   getIsShowDetail,
   refreshNumber,
 } from "./hooks/hooks";
-import { ActionOpenExtensionPreferences } from "./components/action-open-extension-preferences";
+import { ActionOpenCommandPreferences } from "./components/action-open-command-preferences";
 import { ActionCopyFile } from "./components/action-copy-file";
 import { Preferences } from "./types/preferences";
 import { FileContentInfo } from "./types/file-content-info";
@@ -341,7 +341,7 @@ function SendToDirectoryItem(props: {
             />
           </ActionPanel.Section>
 
-          <ActionOpenExtensionPreferences />
+          <ActionOpenCommandPreferences />
         </ActionPanel>
       }
     />